requirements.txt
progressbar~=2.5
torch~=1.5.0
numpy~=1.18.5
Pillow~=7.1.2
opencv-python~=4.2.0.34
scipy~=1.4.1
matplotlib~=3.2.1
scikit-image~=0.17.2
git+https://github.com/pdollar/coco.git#egg=pycocotools&subdirectory=PythonAPI
tensorboardX
jsonschema
torchvision~=0.6.0
progressbar2
skeltorch==1.1.0
jpeg4py~=0.1.4
tqdm~=4.46.1
seaborn~=0.10.1
pandas~=1.0.5
tabulate
moviepy